- rename source files with more meaningful names - update release action - update clang-format
