+++ .github/workflows/release.yml
@@ -1,20 +1,15 @@
 name: Release to NuGet
 
 on:
-  workflow_run:
-    workflows: ["Build Solution"] 
-    types:
-      - completed
+  # workflow_run:
+  #   workflows: ["Build Solution"] 
+  #   types:
+  #     - completed
   push:
     tags:
       - 'v[0-9]+.[0-9]+.[0-9]+*' # Push events to matching v1.0.0 tags
-  pull_request:
-    branches:
-      - master
 
 env:
-  DOTNET_CLI_TELEMETRY_OPTOUT: 1
-  DOTNET_SKIP_FIRST_TIME_EXPERIENCE: 1
   PACKAGE_PROJECT_FILE_PATH: './CwAPI3D.Net.Bridge.sln' 
   NUSPEC_FILE_PATH: './CwAPI3D.Net.Bridge.nuspec'
   BUILD_CONFIGURATION: Release
