--- CwAPI3D.Net.Bridge/src/EBimIfcType.cpp
@@ -1,1 +1,0 @@
-#include "EBimIfcType.h"
