+++ CwAPI3D.Net.Bridge/src/cwapi3d.net.bim-controller.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "CwAPI3D.h"
-#include "IBimController.h"
+#include "cwapi3d.net.ibim-controller.h"
 
 namespace CwAPI3D::Net::Bridge
 {
