+++ CwAPI3D.Net.Bridge/src/cwapi3d.net.bim.ifc.type.h
@@ -35,4 +35,4 @@ public
 		ifcFastener,
 		ifcMechanicalFastener,
 	};
-} // namespace CwAPI3D_Net
+} // namespace CwAPI3D::Net::Bridge
