+++ CwAPI3D.Net.Bridge/src/cwapi3d.net.utils.h
