+++ CwAPI3D.Net.Bridge/src/cwapi3d.net.vector3d.h
